Add files via upload
This script solves for the challenge of identifying the first and last digit in the string using a python script.  You need an input file with the lines of text, an output file that writes the results of the script and a sum file that prints the sum of the 2 digit numbers.  This could likely be a shorter process, however I'm new to python and this is the most efficient way I've figured out how to solve this first challenge.  Hope this helps!