Add files via upload
This is the second step I took in the challenge.  I wanted to create a script that created a dictionary of the words zero through nine so that it would read the file and look for the spelled out words only.  Once I got it to work, I created another script to convert those terms to numbers.  That script is the next script in this repo.